+++ src/cmd/ksh93/include/history.h
@@ -26,7 +26,6 @@
  */
 
 #include	<ast.h>
-#include	<sfio.h>
 
 #define HIST_CHAR	'!'
 #define HIST_VERSION	1		/* history file format version no. */
@@ -36,7 +35,7 @@ typedef struct
 	Sfdisc_t	histdisc;	/* discipline for history */
 	Sfio_t		*histfp;	/* history file stream pointer */
 	char		*histname;	/* name of history file */
-	long		histind;	/* current command number index */
+	int32_t		histind;	/* current command number index */
 	int		histsize;	/* number of accessible history lines */
 #ifdef _HIST_PRIVATE
 	_HIST_PRIVATE
